os/bluestore: split memory 50/50 between bluestore and rocksdb
We don't yet have a clear answer on where our memory is best spent once
we are above kv_min.  Previous testing showed that bluestore was a better
bet, but that predates the enabling of bloom filters and/or inclusion of
bloom filters in the rocksdb cache size.

A 50/50 split hedges out bets until we have more information.  At worst
we are misusing half of any addition memory given to the OSD (vs
potentially more than that for any config that steers more memory one way
or the other).

We still spend no memory on data currently.  That should probably change.

Signed-off-by: Sage Weil <sage@redhat.com>