mgr: add mgr daemon to DaemonStateIndex with metadata (hostname)
This commit changes when (and how) the mgr daemons are added to the
DaemonStateIndex. A change in the mgr map now doesn't immediately
trigger the addition of a mgr daemon, but triggers a MetadataUpdate
operation so that the mgr daemon is added with metadata queried from the
mon (most notably the hostname).
Fixes: http://tracker.ceph.com/issues/23286

Signed-off-by: Jan Fajerski <jfajerski@suse.com>