fuse: handle errors appropriately when getting group list
Currently, the GET_GROUPS macro just ignores errors that occur when
fetching the current task's groups. The callers then end up casting
the error to a count.

Change the macro to a real function, and just have it cleanly ignore
errors when getting groups fails. We do log the error to the debug
log however in that case.

Tracker: http://tracker.ceph.com/issues/23446
Signed-off-by: Jeff Layton <jlayton@redhat.com>