rbd: add notrim option to rbd map
Add the notrim option to the rbd map command to avoid
discarding allocated blocks from a fully provisioned image,
and add the description of notrim option to manpage.

When enabled, user programs such as mkfs.xfs will receive -EOPNOTSUPP
if it sends a discard request for an image with this option.

Signed-off-by: Hitoshi Kamei <hitoshi.kamei.xm@hitachi.com>
Cc: Mitsuo Hayasaka <mitsuo.hayasaka.hu@hitachi.com>