common/hobject: compare two objects' key directly
The original implement does not appear to be a problem for two different
objects which are usually of different oids/names.
However, for a specific head object and its posterity snap objects, this
is not very effective since we are going to compare oids/names twice
(because they are definitely equivalent).

Signed-off-by: xie xingguo <xie.xingguo@zte.com.cn>