common/Formatter: fix string_view usage for {json,xml}_stream_escaper
These are passing a simple const char* without a length, which breaks for
binary strings with \0 in them.

Broken by dc79eab7a12a63f656802912cac1fcf191b67f5b.

This code is still a mess because there is a mix of boost::string_view
and std::string_view throughout the code.  A cleanup is in order!

Signed-off-by: Sage Weil <sage@redhat.com>