mon/LogMonitor: simplify summary logic
Instead of sending the last 1 message, send from the last committed
segment.

This may be >1 entry, which is a bit odd, but otherwise the behavior is
similar.  Note that since the channel filtering is happening on the
client this may mean a higher chance of getting something visible at
all.

Signed-off-by: Sage Weil <sage@redhat.com>