osd/ECBackend: wait for apply for luminous peers
Avoid completing a write until luminous peers also apply the change.
This is overkill, but works around a problem where completing our write
here allows the next read to start too early.  This is because the
PrimaryLogPG is handling the write-to-read ordering, but completing the
write releases the write lock in the PG.

Signed-off-by: Sage Weil <sage@redhat.com>