build: Allow use of Clang to build Python extensions under Fedora 28
Fedora builds Python with hardening flags that are unsupported by
Clang. Python insists on building extensions with all the flags it was
given, not just ones relevant to finding Python libraries and
includes.

Ended up monkey patching customize_compiler to pull out the offending
flags when Clang is in use.

Signed-off-by: Adam C. Emerson <aemerson@redhat.com>