mgr/telegraf: Use generators in Python to streamline output
Instead of collection all data in a list prior to sending use
generators inside Python to send out data as soon as the Mgr
has supplied us with the data.

On larger systems with thousands of OSDs this can reduce the
amount of time it takes to collect and send the data.

Telegraf also likes this more as we send less data in one go.

Signed-off-by: Wido den Hollander <wido@42on.com>