messages/MOSDPGLog: encode epoch for query_epoch for pre-nautilus OSDs
This fix is analogous to 98e1d365ea5cfdc777673ee6111f9c6ca7223ec6.  As of
65baaf7478fc0808032653a2fda71774ab18ea14 we set query_epoch to a lower
value reliably, but we need to encode the latest epoch for pre-nautilus
OSDs because they do not set last_peering_epoch reliably and havin a lower
query_epoch will cause them to discard this message.

Signed-off-by: Sage Weil <sage@redhat.com>