mgr/dashboard: Improve SummaryService and TaskWrapperService
When you create a new task, and it stays running, it will be added automatically
to the summary data.

This will allows for us to deal with it more quickly, by subscribing
to the summaryService, and removes the need to pass a runningTasks array
between services.

Added 3 new methods to SummaryService.

Signed-off-by: Tiago Melo <tmelo@suse.com>