doc: explain how to mkdir under .snap
Prior to this change, the CephFS snapshot instructions were not clear
whether users should "mkdir" the ".snap" directory or not.

Update the documentation to clarify that users should make a directory
under .snap with a name of their choice.

Thanks to John Spray <jspray@redhat.com> on ceph-users for this
information.

Signed-off-by: Ken Dreyer <kdreyer@redhat.com>