mds: cleanup nested auth pins tracking
Limit nested auth pins tracking within dirfrag. This is preparation for
multi-threads (different threads modify different dirfrags)

The main change is that, when freezing a subtree, we need to traverse
the subtree, manually mark dirfrags as 'freezing'. Compare to the old
approach, more work need to be done in freeze/unfreeze tree functions.
But the new approach simplifies checks for freezing/frozen object. They
do not need to walk up the tree to find subtree root.

Signed-off-by: "Yan, Zheng" <zyan@redhat.com>