mgr/balancer: add crush_compat_metrics param
Add a crush_compat_metrics options to allow choosing between pgs,
objects, and bytes when rebalancing with crush-compat mode.

Multi-values are supported, so that the placement scoring can be
a function of several metrics. By default we score based on the
mean of pgs, objects, and bytes scores for all roots.

Signed-off-by: Dan van der Ster <daniel.vanderster@cern.ch>