crushtool: implement --reclassify
Add two modes of reclassification of existing hierarchies:

--classify-root <rootname> <class> will rewrite a hierarchy from an
existing root so that all of its devices are a different class.  Rules
that reference that root will be implicitly adjusted to 'take <rootname>
class <class>'. Ids will be preserved.

--classify-bucket <match> <class> <default-parent> will match a pattern
in the bucket name, where % at the beginning or end of the string is used
as a wildcard (e.g., "%-ssd" will match an "-ssd" suffix, "foo-%" will
match a "foo-" prefix).  Each such bucket is mapped to a "base" bucket
(with the suffix or prefix), items are labeled with the appropriate
class and the moved to that base bucket, and rules adjusted.  The
<default-parent> is used as the parent if the base bucket doesn't exist
and has to be created.

Similarly,
--classify-bucket <bucket> <class> <base-bucket> does the same but
a single existing bucket is mapped to an existing base bucket.  For
example, there is often an 'ssd' bucket that is the counterpart for
the 'default' root; '--classify-bucket ssd ssd default' will map it over.

Signed-off-by: Sage Weil <sage@redhat.com>