pybind/ceph_argparse.py: do not create file for validating CephFilepath
before this change, a file is opened with append mode for validating the
specified CephFilepath argument. but this file is never removed after
the validation. we cannot assume that this file will be overwritten or
removed by the Ceph daemon who serves the command.

after this change, no file is created for the validation, instead we
check if the file is readable or writable.

Signed-off-by: Kefu Chai <kchai@redhat.com>