mgr: race between daemon state and service map in 'service status'
If a daemon is removed from daemon_state it is still present in the
service map. So when 'service status' iterates daemons from service map,
it may not find it in daemon_state. Use pending_service_map instead
when building daemons lists, which is up ot date with daemon_state.

Fixes: http://tracker.ceph.com/issues/36656
Signed-off-by: Mykola Golub <mgolub@suse.com>