rgw: change the way sysobj filters raw attributes, fix bucket sync state xattrs
Fixes: http://tracker.ceph.com/issues/37281

This is a regression, introduced in 00a0bde10ce18ef448825ad2e5ef91065f9cc16f.
The problem was that the original change fixed was that we didn't filter
out sys object attributes that needed to be filtered (e.g., ver objclass
private attributes that were set on the object). We shouldn't know about
these attributes. RGW mostly prefixes all its attributes with "user.rgw."
so in most cases just filtering out everything that doesn't start with
that is fine. However, prior to refactoring, we had multiple ways to read
sys object, and apparently there was at least one case where we didn't
prefix the attrs correctly (bucket sync state).

This commit adds the option to read all of sys object's xattrs, and uses
that in the case where we needed to do it. However, it also modifies the
bucket sync state xattrs to be prefixed (in a backward compatible way).

Signed-off-by: Yehuda Sadeh <yehuda@redhat.com>