mgr/dashboard: Notification queue
It's now possible to send an array of notification configs to the notification
service. These array is cached for, by default 500ms, before this array
is unified and than converted to full notifications that are seen by the
user.

Useful if different service can come to the same conclusions regarding
the notification. This is the case for the prometheus notification service
and the prometheus alert service.

Fixes: https://tracker.ceph.com/issues/37469
Signed-off-by: Stephan Müller <smueller@suse.com>