buffer: fix the traits of list::iterator
before this change, the constness of value, pointer, etc traits of
list::iterator was wrong.  also, because std::iterator is deprecated
in C++17, we need to define the traits manually. so, do this change.

these traits could be potentially used anywhere in the source tree.
but a noteworthy user is is_const_iterator<> in denc.h.

Signed-off-by: Kefu Chai <kchai@redhat.com>