mgr/dashboard: CdNotificationConfig
This model will now be taken as single argument in the 'show' method
of the notification service. Through using this model it's easy to
define notifications before sending them out. Which can be helpful
if you have to process an array of possible notification and the
possibility that those notifications would look a like.

This scenario doesn't exist in the current code base, but it's the
case for new feature to show alerts from Prometheus.

Fixes: https://tracker.ceph.com/issues/37469
Signed-off-by: Stephan Müller <smueller@suse.com>