mgr/dashboard: Generate Open API Spec info
Fixes: http://tracker.ceph.com/issues/24763

Makes it possible for developers to provide information of
endpoints and controllers in decorators.
If information is provided it is transposed to fit Open
API Specification and displayed on Swagger UI page.

Features:
Possible to generate group (=tag) descriptions for controllers.
Possible to assign controller to another group.
Possible to generate descriptions for endpoints.
Possible to assign endpoints to another group.
Possible to generate type + description for responses.
Possible to generate type + description for path/query parameter.
Possible to generate type + description for request body parameters.
Body parameters and responses can be primitive types, dictionaries,
lists of primitive types and dictionaries in dictionaries, lists or
tuples, as well as combinations of these types.

Signed-off-by: Tina Kallio <tina.kallio@gmail.com>